新写法Begin

+++ Bbs/View/index.php
@@ -17,9 +17,11 @@
                     </a>
                 </td>
                 <td>
-                <span class="truncate tooltipped" data-position="bottom" data-delay="50" data-tooltip="<?php echo T($row['tips']); ?>">
-                    <?php echo T($row['tips']); ?>
-                </span>
+                    <?php if (!empty($row['tips'])): ?>
+                        <span class="truncate tooltipped" data-position="bottom" data-delay="50" data-tooltip="<?php echo T($row['tips']); ?>">
+                            <?php echo T($row['tips']); ?>
+                        </span>
+                    <?php endif; ?>
                 </td>
             </tr>
         <?php endforeach; ?>
